[Refactor] DataSource를 관리하는 설정파일을 외부 클래스로 분리
